Title: 바구니 안의 사탕, Time: 1002ms, Memory: 87MB, Status: Wrong Answer - CodeTree

+++ 240422/바구니 안의 사탕/candy-in-the-basket.java
@@ -3,7 +3,7 @@ import java.util.*;
 public class Main {
     public static int[] arr = new int[1000001];
     public static int maxLength;
-    public static int maxValue = Integer.MIN_VALUE;
+    public static long maxValue = Long.MIN_VALUE;
 
     public static void find(){
